+++ src/components/Form/FormComboboxSingle.tsx
@@ -1,8 +1,8 @@
 import { Combobox } from '@headlessui/react';
-import { useEffect, useState } from 'react';
-import { FieldValues, useFormContext } from 'react-hook-form';
-import { ComboboxProps } from './types';
+import { FieldValues } from 'react-hook-form';
 import { GenericDataType } from '../../common';
+import { ComboboxProps } from './types';
+import { useSingleSelectFormSync } from './useSingleSelectFormSync';
 
 export const ComboboxSingle = <
   DataItem extends GenericDataType,
@@ -10,30 +10,28 @@ export const ComboboxSingle = <
 >({
   children,
   className,
-  getItemValue,
+  disabled,
+  formValueMode,
   name,
+  options,
   selectedItems,
   setShowDropdown,
   setSelectedItems,
 }: ComboboxProps<DataItem, Values>): JSX.Element => {
-  const [shouldTouch, setShouldTouch] = useState(false);
-  const { setValue } = useFormContext();
+  useSingleSelectFormSync({
+    name,
+    onItemSelect: () => setShowDropdown(false),
+    options,
+    selectedItems,
+    setSelectedItems,
+    valueMode: formValueMode ?? 'id',
+  });
   const selectedItem = selectedItems[0] ?? null;
-  useEffect(() => {
-    const itemValue =
-      getItemValue !== undefined
-        ? selectedItem !== null
-          ? getItemValue(selectedItem)
-          : ''
-        : selectedItem;
-    setValue<string>(name, itemValue, { shouldDirty: true, shouldTouch });
-    setShowDropdown(false);
-    setShouldTouch(true);
-  }, [selectedItem]);
   return (
     <Combobox
       as="div"
       className={className}
+      disabled={disabled}
       name={name}
       nullable
       onChange={value => value && setSelectedItems([value])}