+++ src/components/Form/types.ts
@@ -1,8 +1,9 @@
+import { Dictionary } from 'lodash';
 import { Dispatch, ReactNode, SetStateAction } from 'react';
 import { FieldValues, Path, UseControllerProps } from 'react-hook-form';
+import { GenericDataType } from '../../common';
 import { SelectProps } from '../Select';
 import { TypeaheadSelectProps } from '../TypeaheadSelect';
-import { GenericDataType } from '../../common';
 
 export interface FormFieldProps<Values extends FieldValues> {
   controllerProps?: Omit<UseControllerProps<Values>, 'name'>;
@@ -18,9 +19,11 @@ export interface ComboboxProps<
   Values extends FieldValues,
 > extends Pick<
     TypeaheadSelectProps<DataItem, Values>,
-    'className' | 'getItemValue' | 'name'
+    'className' | 'formValueMode' | 'name'
   > {
   children: ReactNode;
+  disabled?: boolean;
+  options: Dictionary<DataItem[]>;
   selectedItems: DataItem[];
   setShowDropdown: Dispatch<SetStateAction<boolean>>;
   setSelectedItems: Dispatch<SetStateAction<DataItem[]>>;
@@ -31,10 +34,25 @@ export interface ListboxProps<
   Values extends FieldValues,
 > extends Pick<
     SelectProps<DataItem, Values>,
-    'className' | 'getItemValue' | 'name'
+    'className' | 'formValueMode' | 'name'
   > {
   children?: ReactNode;
   disabled?: boolean;
+  options: Dictionary<DataItem[]>;
   selectedItems: DataItem[];
   setSelectedItems: Dispatch<SetStateAction<DataItem[]>>;
+}
+
+export type FormValueMode = 'item' | 'id';
+
+export interface SelectSyncOptions<
+  Values extends FieldValues,
+  DataItem extends GenericDataType,
+> {
+  name: Path<Values>;
+  onItemSelect?: (items: DataItem[]) => void;
+  options: Dictionary<DataItem[]>;
+  selectedItems: DataItem[];
+  setSelectedItems: Dispatch<SetStateAction<DataItem[]>>;
+  valueMode: FormValueMode;
 }