+++ src/hooks/__tests__/usePrevious.spec.ts
@@ -1,0 +1,35 @@
+import { renderHook } from '@testing-library/react';
+import { usePrevious } from '../usePrevious';
+
+describe('usePrevious', () => {
+  it('should return undefined for the initial value', () => {
+    const { result } = renderHook(() => usePrevious<number>(42));
+    expect(result.current).toBeUndefined();
+  });
+
+  it('should return the previous value after an update', () => {
+    const { result, rerender } = renderHook(
+      ({ value }) => usePrevious<number>(value),
+      {
+        initialProps: { value: 42 },
+      },
+    );
+    expect(result.current).toBeUndefined();
+    rerender({ value: 43 });
+    expect(result.current).toBe(42);
+    rerender({ value: 44 });
+    expect(result.current).toBe(43);
+  });
+
+  it('should not update the previous value if the current value is the same', () => {
+    const { result, rerender } = renderHook(
+      ({ value }) => usePrevious<number>(value),
+      {
+        initialProps: { value: 42 },
+      },
+    );
+    expect(result.current).toBeUndefined();
+    rerender({ value: 42 });
+    expect(result.current).toBeUndefined();
+  });
+});