+++ src/hooks/useTypeaheadQuery.ts
@@ -4,7 +4,6 @@ import { GenericDataType } from '../common';
 
 export interface UseTypeaheadQueryOptions<DataItem> {
   debounceTime?: number;
-  defaultOptions?: DataItem[];
   onDebouncedChange: (value: string) => void;
   options?: DataItem[];
 }